+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 3.62.0 (Unreleased)
 
 ### Added
-- Support Proxy Protocol for `oci_load_balancer_listener`
+- Support for Proxy Protocol for `oci_load_balancer_listener`
+- Support for specifying db version while creating a database for ADB Serverless
 
 ## 3.61.0 (February 05, 2020)